+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 Changes are listed by minor version, from newest to oldest. Under each minor version, patches are listed from oldest to newest.
 
 
+## In Progress
+
+New properties:
+* QElement.center (ElementCenter)
+
+New descriptors:
+* ElementCenter
+
+
 ## 0.3: Relative Positioning
 
 **15 Oct 2014.** QElement's `diff()` method now supports checking relative positions. As well as specifying an absolute position, you can also describe your element relative to other elements. You can also specify an offset if elements aren't exactly aligned: